+++ next/src/routes/_layout/questionnaire.$questionnaireId/_layout-q/releases/index.tsx
@@ -5,9 +5,14 @@ import {
   pendingReleasesQueryOptions,
   releasesQueryOptions,
 } from '@/api/releases'
+import { versionsQueryOptions } from '@/api/versions'
 import ErrorComponent from '@/components/layout/ErrorComponent'
 import ReleaseOverview from '@/components/release/ReleaseOverview'
 import ReleaseOverviewLayout from '@/components/release/ReleasesOverviewLayout'
+import {
+  getLatestVersionId,
+  hasReleaseForVersion,
+} from '@/components/release/overview/utils/utils'
 
 export const Route = createFileRoute(
   '/_layout/questionnaire/$questionnaireId/_layout-q/releases/',
@@ -22,6 +27,7 @@ export const Route = createFileRoute(
     await Promise.all([
       queryClient.ensureQueryData(releasesQueryOptions(questionnaireId)),
       queryClient.ensureQueryData(pendingReleasesQueryOptions(questionnaireId)),
+      queryClient.ensureQueryData(versionsQueryOptions(questionnaireId)),
     ])
   },
 })
@@ -35,9 +41,22 @@ function RouteComponent() {
   const { data: releases } = useSuspenseQuery(
     releasesQueryOptions(questionnaireId),
   )
+  const { data: versions = [] } = useSuspenseQuery(
+    versionsQueryOptions(questionnaireId),
+  )
+
+  const latestVersionId = getLatestVersionId(versions)
+  const isPublishDisabled = hasReleaseForVersion(
+    latestVersionId,
+    releases,
+    requests,
+  )
 
   return (
-    <ReleaseOverviewLayout questionnaireId={questionnaireId}>
+    <ReleaseOverviewLayout
+      questionnaireId={questionnaireId}
+      isPublishDisabled={isPublishDisabled}
+    >
       <ReleaseOverview pendingRequests={requests} releases={releases} />
     </ReleaseOverviewLayout>
   )
@@ -47,7 +66,10 @@ function CustomLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   const { questionnaireId } = Route.useParams()
 
   return (
-    <ReleaseOverviewLayout questionnaireId={questionnaireId}>
+    <ReleaseOverviewLayout
+      questionnaireId={questionnaireId}
+      isPublishDisabled={false}
+    >
       {children}
     </ReleaseOverviewLayout>
   )